refactor(shell): route node and .jsh through realm runtime
Replaces the hand-rolled AsyncFunction + sandbox-iframe paths in
jsh-executor.ts (572→~110 lines) and node-command.ts (616→~85
lines) with a single runInRealm({kind:'js'}) call.

The JS execution engine moves to kernel/realm/js-realm-shared.ts
so both the standalone DedicatedWorker (js-realm-worker.ts) and
the per-task sandbox iframe (sandbox.html) drive the same code
path. Tests get an in-process realm factory (realm-inprocess.ts)
that exercises the same runJsRealm engine without a real
DedicatedWorker.

Headline win: node -e 'while(true){}' is now hard-killable. The
realm runner's SIGKILL hook calls worker.terminate() /
iframe.remove() synchronously, so kill -KILL <pid> exits 137 in
~10 ms regardless of whether the user code yields. SIGINT/SIGTERM
keep their cooperative-record semantic.

PM resolution falls back to globalThis.__slicc_pm so node -e
without an explicit pmConfig still registers under the same
table ps/kill see.

node-command-loadmodule.test.ts moves its assertions from the
legacy node-command.ts wrappedCode template to sandbox.html where
the extension-mode load-module path now lives.
